+++ python/ray/serve/tests/unit/BUILD.bazel
@@ -4,6 +4,7 @@ load("//bazel:python.bzl", "py_test_module_list", "py_test_module_list_with_env_
 py_library(
     name = "conftest",
     srcs = ["conftest.py"],
+    deps = ["//python/ray/serve/tests/unit/common:mock_replica_actor_wrapper"],
 )
 
 py_test_run_all_subdirectory(
@@ -14,6 +15,7 @@ py_test_run_all_subdirectory(
     tags = ["team:serve"],
     deps = [
         ":conftest",
+        "//python/ray/serve/tests/unit/common:mock_replica_actor_wrapper",
         "//python/ray/serve:serve_lib",
         "//python/ray/serve/tests:common",
     ],
@@ -37,6 +39,7 @@ py_test_module_list(
     ],
     deps = [
         ":conftest",
+        "//python/ray/serve/tests/unit/common:mock_replica_actor_wrapper",
         "//python/ray/serve:serve_lib",
         "//python/ray/serve/tests:common",
     ],
@@ -81,6 +84,7 @@ py_test_module_list_with_env_variants(
     ],
     deps = [
         ":conftest",
+        "//python/ray/serve/tests/unit/common:mock_replica_actor_wrapper",
         "//python/ray/serve:serve_lib",
         "//python/ray/serve/tests:common",
     ],
